+++ src/Products/PloneMeeting/browser/viewlets.py
@@ -49,6 +49,28 @@ class PMMeetingBatchActionsViewlet(BatchActionsViewlet):
         return True
 
 
+class AnnexesBatchActionsViewlet(BatchActionsViewlet):
+    """ """
+
+    section = "annexes"
+
+    def available(self):
+        """ """
+        return True
+
+    @property
+    def select_item_name(self):
+        """Manage fact that in the annexes, there are 2 tables
+          (annexes and decision annexes) that use a different name
+          for the checkbox column."""
+        value = None
+        if self.request.get('categorized_tab').portal_type == 'annexDecision':
+            value = "select_item_annex_decision"
+        else:
+            value = super(AnnexesBatchActionsViewlet, self).select_item_name
+        return value
+
+
 class HeldPositionBackRefs(ViewletBase):
     """Display elements using held_position."""