+++ src/Products/PloneMeeting/browser/views.py
@@ -12,9 +12,12 @@ from collective.documentgenerator.helper.archetypes import ATDocumentGenerationH
 from collective.documentgenerator.helper.dexterity import DXDocumentGenerationHelperView
 from collective.eeafaceted.batchactions import _ as _CEBA
 from collective.eeafaceted.batchactions.browser.views import BaseBatchActionForm
+from collective.eeafaceted.batchactions.browser.views import DeleteBatchActionForm
 from collective.eeafaceted.batchactions.utils import listify_uids
 from eea.facetednavigation.interfaces import ICriteria
 from ftw.labels.interfaces import ILabeling
+from imio.actionspanel.interfaces import IContentDeletable
+from imio.annex.browser.views import DownloadAnnexesBatchActionForm
 from imio.helpers.content import uuidToObject
 from imio.helpers.xhtml import addClassToContent
 from imio.helpers.xhtml import CLASS_TO_LAST_CHILDREN_NUMBER_OF_CHARS_DEFAULT
@@ -2086,9 +2089,8 @@ class UpdateLocalRolesBatchActionForm(BaseBatchActionForm):
         self.cfg = self.tool.getMeetingConfig(context)
 
     def available(self):
-        """Hide it on meetings as it uses IMeetingBatchActionsMarker."""
-        return _checkPermission(ManagePortal, self.context) and \
-            not IMeeting.providedBy(self.context)
+        """Only available to Managers."""
+        return _checkPermission(ManagePortal, self.context)
 
     def _apply(self, **data):
         """ """
@@ -2100,6 +2102,41 @@ class UpdateLocalRolesBatchActionForm(BaseBatchActionForm):
                         context=self.request,
                         default="Updated accesses for ${number_of_elements} element(s).")
         api.portal.show_message(msg, request=self.request)
+
+
+class PMDeleteBatchActionForm(DeleteBatchActionForm):
+    """ """
+
+    section = "annexes"
+
+    def __init__(self, context, request):
+        super(PMDeleteBatchActionForm, self).__init__(context, request)
+        self.tool = api.portal.get_tool('portal_plonemeeting')
+        self.cfg = self.tool.getMeetingConfig(context)
+
+    def available(self):
+        """ """
+        return "delete" in self.cfg.getEnabledAnnexesBatchActions() and \
+               _checkPermission(ModifyPortalContent, self.context)
+
+    def get_deletable_elements(self):
+        """ """
+        deletables = [brain for brain in self.brains
+                      if IContentDeletable(brain.getObject()).mayDelete()]
+        return deletables
+
+
+class PMDownloadAnnexesBatchActionForm(DownloadAnnexesBatchActionForm):
+    """ """
+
+    def __init__(self, context, request):
+        super(PMDownloadAnnexesBatchActionForm, self).__init__(context, request)
+        self.tool = api.portal.get_tool('portal_plonemeeting')
+        self.cfg = self.tool.getMeetingConfig(context)
+
+    def available(self):
+        """ """
+        return "download-annexes" in self.cfg.getEnabledAnnexesBatchActions()
 
 
 class DisplayMeetingConfigsOfConfigGroup(BrowserView):