+++ CHANGES.rst
@@ -283,6 +283,13 @@ Changelog
   Also added values `deliberation_motivation` and `deliberation_decision` to
   possible values returned by `ItemDocumentGenerationHelperView.deliberation_for_restapi`.
   [gbastien]
+- Enabled batch actions on annexes:
+
+  - Batch actions `Delete` and `Download as Zip` are available;
+  - Added `MeetingConfig.enabledAnnexesBatchActions` attribute to be able enable
+    or disable batch actions, by default only the `Download` action is enabled.
+
+  [gbastien]
 
 4.2b11 (2021-01-19)
 -------------------